refactor(tensor): prepare QTensor renaming
