+++ quanto/__init__.py
@@ -2,5 +2,5 @@ __version__ = "0.0.11dev"
 
 from .calibrate import *
 from .nn import *
-from .qtensor import *
+from .tensor import *
 from .quantize import *
