+++ quanto/tensor/func.py
