+++ quanto/tensor/ops.py
