+++ test/tensor/test_absmax.py
